🧩 78JKTQ task: Clear cloud hotspot regressions
