+++ scripts/oversized-test-baseline.json
@@ -2,8 +2,8 @@
   "schema_version": 2,
   "threshold_lines": 1000,
   "budgets": {
-    "max_entries": 12,
-    "max_total_lines": 13649,
+    "max_entries": 11,
+    "max_total_lines": 12574,
     "max_file_lines": 1300
   },
   "entries": [
@@ -34,10 +34,6 @@
     {
       "file": "packages/agentplane/src/cli/run-cli.core.test.ts",
       "lines": 1046
-    },
-    {
-      "file": "packages/agentplane/src/backends/task-backend.cloud.test.ts",
-      "lines": 1075
     },
     {
       "file": "packages/agentplane/src/commands/doctor.command.open-pr.test.ts",
